src/requirements.txt
attrs==20.3.0
bcrypt==3.2.0
blinker==1.4
cffi==1.14.5
click==7.1.2
coverage==5.4
Flask==1.1.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.0.2
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
gunicorn==20.0.4
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycparser==2.20
PyJWT==2.0.1
pymongo==3.11.3
pyparsing==2.4.7
pytest==6.2.2
pytest-cov==2.11.1
pytest-flask==1.1.0
python-dotenv==0.15.0
six==1.15.0
toml==0.10.2
Werkzeug==1.0.1
